# [perf] use optimize=True in np.einsum